# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "fabwaste api documentation",
        "description": "API Documentation",
        "contact": {
            "email": "admin@fabwaste.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "view all requests",
                "operationId": "2fb9887cd47a1402dea585c7b53034bc",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/myrequests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "view requests belong to you",
                "operationId": "ba1d34c9d4cc7daf18d29887fd478f57",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/accepted/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "view accepted requests belong to you",
                "operationId": "192f01cd94acea0cf147afb861612cfe",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/request/store/{id}": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "request store",
                "operationId": "3ae4d10f89dab9ea53b3868ccdeb721c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "storeId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/accept/request/{id}": {
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "accept request",
                "operationId": "b1b8be828316ae4a12dd65a2a8da33b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "requestId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reject/request/{id}": {
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "reject request",
                "operationId": "c92540baadda1acf63ca2bcbd2e20c99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "requestId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cancel/request/{id}": {
            "delete": {
                "tags": [
                    "Requests"
                ],
                "summary": "cancel request",
                "operationId": "b00a46bd4ea4c56f5e8517b4f7a2a0be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "requestId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/store/requests": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "view strore requests",
                "operationId": "ee9bdc296080a7d863732b15098c45f3",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}